+++ src/model.py
@@ -8,6 +8,7 @@ from reporter import Reporter
 import ast
 from sklearn.model_selection import GridSearchCV
 import pickle
+import random
 from sklearn.model_selection import LeaveOneGroupOut
 from sklearn.model_selection import StratifiedKFold
 
@@ -22,7 +23,9 @@ class Model:
         self.target = self.util.config_val('DATA', 'target', 'emotion')
         self.run = 0
         self.epoch = 0
-        self.loso = False
+        self.loso = self.util.config_val('MODEL', 'loso', False)
+        self.logo = self.util.config_val('MODEL', 'logo', False)
+        self.xfoldx = self.util.config_val('MODEL', 'k_fold_cross', False)
 
     def set_id(self, run, epoch):
         self.run = run
@@ -35,8 +38,7 @@ class Model:
         annos = self.df_train.append(self.df_test)
         targets = annos[self.target]
         _skf = StratifiedKFold(n_splits=int(self.xfoldx))
-        truths = []
-        preds = []
+        truths, preds, results = [], [], []
         g_index = 0
         # leave-one-speaker loop    
         for train_index, test_index in _skf.split(
@@ -51,6 +53,7 @@ class Model:
             predict_y = self.clf.predict(truth_x)
             report = Reporter(truth_y.astype(float), predict_y, self.run, self.epoch)
             self.util.debug(f'result for fold {g_index}: {report.get_result().get_test_result()} ')
+            results.append(float(report.get_result().test))            
             truths.append(truth_y)
             preds.append(predict_y)
             g_index += 1
@@ -65,16 +68,17 @@ class Model:
         )
         self.truths = truth
         self.preds = pred        
+        results = np.asarray(results)
+        self.util.debug(f'KFOLD: {self.xfoldx} folds: mean {results.mean():.3f}, std: {results.std():.3f}')
 
-    def _logo(self):
+    def _loso(self):
         # ignore train and test sets and do a "leave one speaker out"  evaluation
         self.util.debug('ignoring splits and doing LOSO')
         feats = self.feats_train.df.append(self.feats_test.df)
         annos = self.df_train.append(self.df_test)
         targets = annos[self.target]
         _logo = LeaveOneGroupOut()
-        truths = []
-        preds = []
+        truths, preds, results = [], [], []
         speakers = annos['speaker']
         g_index = 0
         # leave-one-speaker loop    
@@ -90,10 +94,60 @@ class Model:
             truth_x = feats.iloc[test_index].to_numpy()
             truth_y = targets[test_index]
             predict_y = self.clf.predict(truth_x)
-
             report = Reporter(truth_y.astype(float), predict_y, self.run, self.epoch)
             self.util.debug(f'result for speaker {g_index}: {report.get_result().get_test_result()} ')
+            truths.append(truth_y)
+            preds.append(predict_y)
+            g_index += 1
+            results.append(float(report.get_result().test))            
+        # combine speaker folds
+        truth = pd.concat(truths)
+        truth.name = 'truth'
+        pred = pd.Series(
+            np.concatenate(preds),
+            index=truth.index,
+            name='prediction',
+        )
+        self.truths = truth
+        self.preds = pred        
+        results = np.asarray(results)
+        self.util.debug(f'LOSO: {self.loso} folds: mean {results.mean():.3f}, std: {results.std():.3f}')
 
+    def _do_logo(self):
+        # ignore train and test sets and do a "leave one speaker group out"  evaluation
+        self.util.debug('ignoring splits and doing LOSGO')
+        feats = self.feats_train.df.append(self.feats_test.df)
+        annos = self.df_train.append(self.df_test)
+        targets = annos[self.target]
+        _logo = LeaveOneGroupOut()
+        truths, preds, results = [], [], []
+        # get unique list of speakers
+        speakers = annos['speaker'].unique()
+        # create a random dictionary of groups
+        sdict = {}
+        for i, s in enumerate(speakers):
+            sdict[s] = random.sample(range(int(self.logo)), 1)[0]    
+        # add this to the annotations  
+        annos['speaker_groups'] = annos['speaker'].apply(lambda x: str(sdict[x]))
+        speaker_groups = annos['speaker_groups']
+        g_index = 0
+        # leave-one-speaker loop    
+        for train_index, test_index in _logo.split(
+            feats, 
+            targets, 
+            groups=speaker_groups,
+        ):
+            train_x = feats.iloc[train_index].to_numpy()
+            train_y = targets[train_index]
+            self.clf.fit(train_x, train_y)
+            
+            truth_x = feats.iloc[test_index].to_numpy()
+            truth_y = targets[test_index]
+            predict_y = self.clf.predict(truth_x)
+            report = Reporter(truth_y.astype(float), predict_y, self.run, self.epoch)
+            result = report.get_result().get_test_result()
+            self.util.debug(f'result for speaker group {g_index}: {result} ')
+            results.append(float(report.get_result().test))
             truths.append(truth_y)
             preds.append(predict_y)
             g_index += 1
@@ -107,17 +161,21 @@ class Model:
             name='prediction',
         )
         self.truths = truth
-        self.preds = pred        
+        self.preds = pred
+        results = np.asarray(results)
+        self.util.debug(f'LOGO: {self.logo} folds: mean {results.mean():.3f}, std: {results.std():.3f}')
 
     def train(self):
         """Train the model"""
         # first check if leave on  speaker out is wanted
-        self.loso = self.util.config_val('MODEL', 'loso', False)
         if self.loso:
-            self._logo()
+            self._loso()
+            return
+        # then if leave one speaker group out validation is wanted
+        if self.logo:
+            self._do_logo()
             return
         # then if x fold cross validation is wanted
-        self.xfoldx = self.util.config_val('MODEL', 'k_fold_cross', False)
         if self.xfoldx:
             self._x_fold_cross()
             return
@@ -171,7 +229,7 @@ class Model:
         return predictions
 
     def predict(self):
-        if self.loso or self.xfoldx:
+        if self.loso or self.logo or self.xfoldx:
             report = Reporter(self.truths.astype(float), self.preds, self.run, self.epoch)
             return report
         """Predict the whole eval feature set"""